+++ DataTypeTestingTempScript.R
@@ -1,4 +1,4 @@
-data <- data.frame(SaleID=1:5, Colour=c("Red", "Red", "Green", "Green", "Green", "Green", "Red", "Green", "Red", "Green"),
+dtdata <- data.frame(SaleID=1:5, SaleID2=as.character(1:5), Colour=c("Red", "Red", "Green", "Green", "Green", "Green", "Red", "Green", "Red", "Green"),
                    SaleItem=c("Car", "Lorry", "Car", "Train", "Train", "Lorry", "Car", "Train", "Lorry", "Car"),
                    SaleModel=c("CA", "LA", "CB", "TA", "TB", "LB", "CB", "TC", "LD", "CE"),
                    SaleDate=as.Date(c("2018-05-15", "2018-01-23", "2018-09-03", "2017-12-25", "2018-06-28")),
@@ -8,19 +8,177 @@ data <- data.frame(SaleID=1:5, Colour=c("Red", "Red", "Green", "Green", "Green",
                                                   "2016-11-25 18:12:11 UTC")),
                    IsNewCustomer=c(TRUE,FALSE,FALSE,TRUE,FALSE,FALSE,TRUE,FALSE,FALSE,TRUE),
                    SaleQuantity=as.integer(c(1,3,2,1,5,3,1,2,3,2)),
-                   SaleAmount=c(12.1,2.3,5.6,3.7,1.5,1.1,0.2,3.7,2.5,2.9),
+                   SaleAmount=c(12.1,2.333333333,5.6,3.7,1.5,1.1,0.2,3.7,2.5,2.9),
                    stringsAsFactors=FALSE)
 
 # Data Types: integer, numeric, character, logical, date, POSIXct
 # Not supported:  Factors (convert to character), POSIXlt (use POSIXct), complex (who needs these anyway!)
 
-# # MEASURE DATA TYPE CALCULATION AND FORMAT TESTS:  NO FORMAT = as.character()
+library(openxlsx)
+
+testXLSX <- function(pt) {
+  wb <- createWorkbook(creator = Sys.getenv("USERNAME"))
+  addWorksheet(wb, "Data")
+  pt$writeToExcelWorksheet(wb=wb, wsName="Data",
+                           topRowNumber=1, leftMostColumnNumber=1,
+                           applyStyles=TRUE, mapStylesFromCSS=TRUE)
+  saveWorkbook(wb, file="C:\\Users\\Chris\\Desktop\\test.xlsx", overwrite = TRUE)
+}
+
+# ROW/COLUMN DATA TYPE FORMAT TESTS:  NO FORMAT = as.character()
+
+# integer
+
+library(pivottabler)
+pt <- PivotTable$new(processingLibrary="data.table", evaluationMode="batch")
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("SaleQuantity")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# numeric
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("SaleAmount")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# logical
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("IsNewCustomer")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# Date
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("SaleDate")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# POSIXct
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("SaleDT")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# ROW/COLUMN DATA TYPE FORMAT TESTS:  SPRINTF()
+
+# character - N/A
+# date - N/A
+# POSIXct - N/A
 
 # integer
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("SaleQuantity", dataFormat="%i")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# numeric
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("SaleAmount", dataFormat="%.1f")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# logical
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("IsNewCustomer", dataFormat="%i")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# ROW/COLUMN DATA TYPE FORMAT TESTS:  FORMAT()
+
+# integer - N/A
+# character - N/A
+
+# numeric
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("SaleAmount", dataFormat=list(digits=4, nsmall=2))
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# logical
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("IsNewCustomer", dataFormat=c("Existing Customer", "New Customer"))
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# Date
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("SaleDate", dataFormat="%d %b %Y")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+# POSIXct
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
+pt$addColumnDataGroups("SaleID")
+pt$addRowDataGroups("SaleDT", dataFormat="%d %b %Y %H:%M:%S")
+pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
+pt$evaluatePivot()
+pt$renderPivot()
+
+
+
+# MEASURE DATA TYPE CALCULATION AND FORMAT TESTS:  NO FORMAT = as.character()
+
+# integer
+
+library(pivottabler)
+pt <- PivotTable$new()
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
@@ -34,7 +192,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -48,7 +206,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -61,7 +219,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -74,7 +232,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -87,7 +245,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -99,7 +257,6 @@ pt$renderPivot()
 
 
 # MEASURE DATA TYPE FORMAT TESTS:  SPRINTF()
-# = format using as.character
 
 # character - N/A
 # date - N/A
@@ -109,7 +266,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", format="%i", caption="Volume Sold")
@@ -123,7 +280,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -137,7 +294,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -156,7 +313,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="VolumeSold", summariseExpression="sum(SaleQuantity, na.rm=TRUE)", caption="Volume Sold")
@@ -171,7 +328,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -186,7 +343,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -201,7 +358,7 @@ pt$renderPivot()
 
 library(pivottabler)
 pt <- PivotTable$new()
-pt$addData(data)
+pt$addData(dtdata)
 pt$addColumnDataGroups("Colour")
 pt$addRowDataGroups("SaleItem")
 pt$defineCalculation(calculationName="SaleCount", summariseExpression="n()", caption="Sale Count")
@@ -217,8 +374,6 @@ pt$renderPivot()
 
 # all of the above examples use the default formatting option (of as.character()) since no format parameter is specified.
 # need to:
-#   - write tests that use the format parameter and check they work...
-#     ... both for pivot rows/columns of different types (not yet done) and measures of different types (done, see above)
 #   - check export to Excel to see how that handles exporting each of the above
 #   - check latex export for the same reason
 #   - write automated test cases for the above